+++ src/main/java/kitchenpos/order/eatin/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatin.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
