+++ src/main/java/kitchenpos/order/eatin/infra/JpaOrderRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.order.infra;
+package kitchenpos.order.eatin.infra;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.eatin.domain.Order;
+import kitchenpos.order.eatin.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
