+++ src/main/java/kitchenpos/ordertable/application/OrderTableService.java
@@ -1,7 +1,7 @@
 package kitchenpos.ordertable.application;
 
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.eatin.domain.OrderRepository;
+import kitchenpos.order.eatin.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
