+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,10 +3,10 @@ package kitchenpos;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderType;
+import kitchenpos.order.eatin.domain.Order;
+import kitchenpos.order.eatin.domain.OrderLineItem;
+import kitchenpos.order.eatin.domain.OrderStatus;
+import kitchenpos.order.eatin.domain.OrderType;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
